+++ share/templates/js-lib/gruntfile.js
@@ -29,9 +29,9 @@ module.exports = function(grunt) {
       },
 
       commonjs: {
-        src: ['lib/*.jsx'],
+        src: ['src/*.jsx'],
         output_rule: {
-            regexp: /lib\/(.+)\.jsx/,
+            regexp: /src\/(.+)\.jsx/,
             replace: 'dest\/$1.common.js'
         },
         add_search_path: ['<%= libDir %>'],
@@ -41,9 +41,9 @@ module.exports = function(grunt) {
       },
 
       amd: {
-        src: ['lib/*.jsx'],
+        src: ['src/*.jsx'],
         output_rule: {
-            regexp: /lib\/(.+)\.jsx/,
+            regexp: /src\/(.+)\.jsx/,
             replace: 'dest\/$1.amd.js'
         },
         add_search_path: ['<%= libDir %>'],
@@ -53,9 +53,9 @@ module.exports = function(grunt) {
       },
 
       closure: {
-        src: ['lib/*.jsx'],
+        src: ['src/*.jsx'],
         output_rule: {
-            regexp: /lib\/(.+)\.jsx/,
+            regexp: /src\/(.+)\.jsx/,
             replace: 'dest\/$1.closure.js'
         },
         add_search_path: ['<%= libDir %>'],
@@ -65,9 +65,9 @@ module.exports = function(grunt) {
       },
 
       global: {
-        src: ['lib/*.jsx'],
+        src: ['src/*.jsx'],
         output_rule: {
-            regexp: /lib\/(.+)\.jsx/,
+            regexp: /src\/(.+)\.jsx/,
             replace: 'dest\/$1.global.js'
         },
         add_search_path: ['<%= libDir %>'],
@@ -77,9 +77,9 @@ module.exports = function(grunt) {
       },
 
       standard: {
-        src: ['lib/*.jsx'],
+        src: ['src/*.jsx'],
         output_rule: {
-            regexp: /lib\/(.+)\.jsx/,
+            regexp: /src\/(.+)\.jsx/,
             replace: 'dest\/$1.js'
         },
         add_search_path: ['<%= libDir %>'],
@@ -94,7 +94,7 @@ module.exports = function(grunt) {
       },
 
       doc: {
-        src: ['<%= libDir %>/*.jsx'],
+        src: ['<%= libDir %>/*.jsx', 'src/*.jsx'],
         add_search_path: ['<%= libDir %>'],
         dest: '<%= docDir %>',
         mode: 'doc'