+++ share/templates/node-app/src/__name__.jsx
@@ -1,5 +1,5 @@
 /**
- * {{& OriginalName }}
+ * {{& Name }}
  *
  * {{& description }}
  *