+++ share/templates/web-canvas-app/gruntfile.js
@@ -34,7 +34,7 @@ module.exports = function(grunt) {
     jsx: {
       build: {
         src: ['<%= srcDir %>/*.jsx'],
-        add_search_path: ['<%= libDir %>', 'node_modules/*/src'],
+        add_search_path: ['<%= libDir %>'],
         dest: '<%= buildDir %>/',
         executable: 'web',
         release: true
@@ -42,12 +42,12 @@ module.exports = function(grunt) {
 
       test: {
         src: ['<%= testDir %>/*.jsx'],
-        add_search_path: ['<%= libDir %>', '<%= srcDir %>', 'node_modules/*/src'],
+        add_search_path: ['<%= libDir %>', '<%= srcDir %>'],
         test: true
       },
 
       doc: {
-        src: ['<%= libDir %>/*.jsx', '<%= srcDir %>/*.jsx', 'node_modules/*/src'],
+        src: ['<%= libDir %>/*.jsx', '<%= srcDir %>/*.jsx'],
         add_search_path: ['<%= libDir %>', '<%= srcDir %>'],
         dest: '<%= docDir %>',
         mode: 'doc'